Extract serialization logic to standalone class
