+++ src/onepasswordconnectsdk/client.py
@@ -1,36 +1,30 @@
 """Python Client for connecting to 1Password Connect"""
 import httpx
 from httpx import HTTPError
-from dateutil.parser import parse
 import json
 import os
-import re
-import datetime
-import onepasswordconnectsdk
+
+from onepasswordconnectsdk.serializer import Serializer
+from onepasswordconnectsdk.utils import is_valid_uuid
+from onepasswordconnectsdk.errors import (
+    FailedToRetrieveItemException,
+    FailedToRetrieveVaultException,
+    EnvironmentHostNotSetException,
+    EnvironmentTokenNotSetException,
+)
 from onepasswordconnectsdk.models import Item, ItemVault
 from onepasswordconnectsdk.models.constants import CONNECT_HOST_ENV_VARIABLE
 
 ENV_SERVICE_ACCOUNT_JWT_VARIABLE = "OP_CONNECT_TOKEN"
-UUIDLength = 26
 
 
 class Client:
-    PRIMITIVE_TYPES = (float, bool, bytes, str, int)
-    NATIVE_TYPES_MAPPING = {
-        "int": int,
-        "float": float,
-        "str": str,
-        "bool": bool,
-        "date": datetime.date,
-        "datetime": datetime.datetime,
-        "object": object,
-    }
-
     """Python Client Class"""
 
-    def __init__(self, session):
+    def __init__(self, session, serializer):
         """Initialize client"""
         self.session = session
+        self.serializer = serializer
 
     def __del__(self):
         self.session.close()
@@ -45,7 +39,7 @@ class Client:
                 f"Unable to retrieve item. Received {response.status_code}\
                                      for {url} with message: {response.json().get('message')}"
             )
-        return self.deserialize(response.content, "File")
+        return self.serializer.deserialize(response.content, "File")
 
     def get_files(self, item_id: str, vault_id: str):
         url = f"/v1/vaults/{vault_id}/items/{item_id}/files"
@@ -58,7 +52,7 @@ class Client:
                 f"Unable to retrieve item. Received {response.status_code}\
                              for {url} with message: {response.json().get('message')}"
             )
-        return self.deserialize(response.content, "list[File]")
+        return self.serializer.deserialize(response.content, "list[File]")
 
     def get_file_content(self, file_id: str, item_id: str, vault_id: str):
         url = f"/v1/vaults/{vault_id}/items/{item_id}/files/{file_id}/content"
@@ -99,10 +93,10 @@ class Client:
         """
 
         vault_id = vault
-        if not self._is_valid_UUID(vault):
+        if not is_valid_uuid(vault):
             vault_id = self.get_vault_by_title(vault).id
 
-        if self._is_valid_UUID(item):
+        if is_valid_uuid(item):
             return self.get_item_by_id(item, vault_id)
         else:
             return self.get_item_by_title(item, vault_id)
@@ -131,7 +125,7 @@ class Client:
                 f"Unable to retrieve item. Received {response.status_code}\
                      for {url} with message: {response.json().get('message')}"
             )
-        return self.deserialize(response.content, "Item")
+        return self.serializer.deserialize(response.content, "Item")
 
     def get_item_by_title(self, title: str, vault_id: str):
         """Get a specific item by title
@@ -165,7 +159,7 @@ class Client:
                     title {title}"
             )
 
-        item_summary = self.deserialize(response.content, "list[SummaryItem]")[0]
+        item_summary = self.serializer.deserialize(response.content, "list[SummaryItem]")[0]
         return self.get_item_by_id(item_summary.id, vault_id)
 
     def get_items(self, vault_id: str):
@@ -192,7 +186,7 @@ class Client:
                      for {url} with message: {response.json().get('message')}"
             )
 
-        return self.deserialize(response.content, "list[SummaryItem]")
+        return self.serializer.deserialize(response.content, "list[SummaryItem]")
 
     def delete_item(self, item_id: str, vault_id: str):
         """Deletes a specified item from a specified vault
@@ -242,7 +236,7 @@ class Client:
                 f"Unable to post item. Received {response.status_code}\
                     for {url} with message: {response.json().get('message')}"
             )
-        return self.deserialize(response.content, "Item")
+        return self.serializer.deserialize(response.content, "Item")
 
     def update_item(self, item_uuid: str, vault_id: str, item: Item):
         """Update the specified item at the specified vault.
@@ -271,7 +265,7 @@ class Client:
                 f"Unable to post item. Received {response.status_code}\
                     for {url} with message: {response.json().get('message')}"
             )
-        return self.deserialize(response.content, "Item")
+        return self.serializer.deserialize(response.content, "Item")
 
     def get_vault(self, vault_id: str):
         """Returns the vault with the given vault_id
@@ -296,7 +290,7 @@ class Client:
                      for {url} with message {response.json().get('message')}"
             )
 
-        return self.deserialize(response.content, "Vault")
+        return self.serializer.deserialize(response.content, "Vault")
 
     def get_vault_by_title(self, name: str):
         """Returns the vault with the given name
@@ -329,7 +323,7 @@ class Client:
                     name {name}"
             )
 
-        return self.deserialize(response.content, "list[Vault]")[0]
+        return self.serializer.deserialize(response.content, "list[Vault]")[0]
 
     def get_vaults(self):
         """Returns all vaults for service account set in client
@@ -352,7 +346,7 @@ class Client:
                      for {url} with message {response.json().get('message')}"
             )
 
-        return self.deserialize(response.content, "list[Vault]")
+        return self.serializer.deserialize(response.content, "list[Vault]")
 
     def build_request(self, method: str, path: str, body=None):
         """Builds a http request
@@ -366,212 +360,11 @@ class Client:
         """
 
         if body:
-            serialized_body = json.dumps(self.sanitize_for_serialization(body))
+            serialized_body = json.dumps(self.serializer.sanitize_for_serialization(body))
             response = self.session.request(method, path, data=serialized_body)
         else:
             response = self.session.request(method, path)
         return response
-
-    def deserialize(self, response, response_type):
-        """Deserializes response into an object.
-
-        :param response: RESTResponse object to be deserialized.
-        :param response_type: class literal for
-            deserialized object, or string of class name.
-
-        :return: deserialized object.
-        """
-        # fetch data from response object
-        try:
-            data = json.loads(response)
-        except ValueError:
-            data = response
-
-        return self.__deserialize(data, response_type)
-
-    def sanitize_for_serialization(self, obj):
-        """Builds a JSON POST object.
-
-        If obj is None, return None.
-        If obj is str, int, long, float, bool, return directly.
-        If obj is datetime.datetime, datetime.date convert to string
-        in iso8601 format.
-        If obj is list, sanitize each element in the list.
-        If obj is dict, return the dict.
-        If obj is OpenAPI model, return the properties dict.
-
-        :param obj: The data to serialize.
-        :return: The serialized form of data.
-        """
-        if obj is None:
-            return None
-        elif isinstance(obj, self.PRIMITIVE_TYPES):
-            return obj
-        elif isinstance(obj, list):
-            return [self.sanitize_for_serialization(sub_obj) for sub_obj in obj]  # noqa: E501
-        elif isinstance(obj, tuple):
-            return tuple(self.sanitize_for_serialization(sub_obj) for sub_obj in obj)  # noqa: E501
-        elif isinstance(obj, (datetime.datetime, datetime.date)):
-            return obj.isoformat()
-
-        if isinstance(obj, dict):
-            obj_dict = obj
-        else:
-            # Convert model obj to dict except
-            # attributes `openapi_types`, `attribute_map`
-            # and attributes which value is not None.
-            # Convert attribute name to json key in
-            # model definition for request.
-            obj_dict = {
-                obj.attribute_map[attr]: getattr(obj, attr)
-                for attr in obj.openapi_types.keys()
-                if getattr(obj, attr) is not None
-            }
-
-        return {
-            key: self.sanitize_for_serialization(val)
-            for key, val in obj_dict.items()
-        }
-
-    def __deserialize(self, data, klass):
-        """Deserializes dict, list, str into an object.
-
-        :param data: dict, list or str.
-        :param klass: class literal, or string of class name.
-
-        :return: object.
-        """
-        if data is None:
-            return None
-
-        if type(klass) == str:
-            if klass.startswith("list["):
-                sub_kls = re.match(r"list\[(.*)\]", klass).group(1)
-                return [self.__deserialize(sub_data, sub_kls) for sub_data in data]  # noqa: E501
-
-            if klass.startswith("dict("):
-                sub_kls = re.match(r"dict\(([^,]*), (.*)\)", klass).group(2)
-                return {
-                    k: self.__deserialize(v, sub_kls) for k, v in data.items()  # noqa: E501
-                }
-
-            # convert str to class
-            if klass in self.NATIVE_TYPES_MAPPING:
-                klass = self.NATIVE_TYPES_MAPPING[klass]
-            else:
-                klass = getattr(onepasswordconnectsdk.models, klass)
-
-        if klass in self.PRIMITIVE_TYPES:
-            return self.__deserialize_primitive(data, klass)
-        elif klass == object:
-            return self.__deserialize_object(data)
-        elif klass == datetime.date:
-            return self.__deserialize_date(data)
-        elif klass == datetime.datetime:
-            return self.__deserialize_datetime(data)
-        else:
-            return self.__deserialize_model(data, klass)
-
-    def __deserialize_primitive(self, data, klass):
-        """Deserializes string to primitive type.
-
-        :param data: str.
-        :param klass: class literal.
-
-        :return: int, long, float, str, bool.
-        """
-        try:
-            return klass(data)
-        except UnicodeEncodeError:
-            return str(data)
-        except TypeError:
-            return data
-
-    def __deserialize_object(self, value):
-        """Return an original value.
-
-        :return: object.
-        """
-        return value
-
-    def __deserialize_date(self, string):
-        """Deserializes string to date.
-
-        :param string: str.
-        :return: date.
-        """
-        try:
-            return parse(string).date()
-        except ImportError:
-            return string
-        except ValueError:
-            raise FailedToDeserializeException(
-                f'Failed to parse `{0}`\
-                 as date object".format(string)'
-            )
-
-    def __deserialize_datetime(self, string):
-        """Deserializes string to datetime.
-
-        The string should be in iso8601 datetime format.
-
-        :param string: str.
-        :return: datetime.
-        """
-        try:
-            return parse(string)
-        except ImportError:
-            return string
-        except ValueError:
-            raise FailedToDeserializeException(
-                f'Failed to parse `{0}`\
-                 as date object".format(string)'
-            )
-
-    def __deserialize_model(self, data, klass):
-        """Deserializes list or dict to model.
-
-        :param data: dict, list.
-        :param klass: class literal.
-        :return: model object.
-        """
-        has_discriminator = False
-        if (
-                hasattr(klass, "get_real_child_model")
-                and klass.discriminator_value_class_map
-        ):
-            has_discriminator = True
-
-        if not klass.openapi_types and has_discriminator is False:
-            return data
-
-        kwargs = {}
-        if (
-                data is not None
-                and klass.openapi_types is not None
-                and isinstance(data, (list, dict))
-        ):
-            for attr, attr_type in klass.openapi_types.items():
-                if klass.attribute_map[attr] in data:
-                    value = data[klass.attribute_map[attr]]
-                    kwargs[attr] = self.__deserialize(value, attr_type)
-
-        instance = klass(**kwargs)
-
-        if has_discriminator:
-            klass_name = instance.get_real_child_model(data)
-            if klass_name:
-                instance = self.__deserialize(data, klass_name)
-        return instance
-
-    def _is_valid_UUID(self, uuid):
-        if len(uuid) is not UUIDLength:
-            return False
-        for c in uuid:
-            valid = (c >= 'a' and c <= 'z') or (c >= '0' and c <= '9')
-            if valid is False:
-                return False
-        return True
 
 
 class AsyncClient:
@@ -597,12 +390,13 @@ def new_client(url: str, token: str, is_async: bool = False):
     Client: The 1Password Connect client
     """
     headers = build_headers(token)
+    serializer = Serializer()
     if is_async:
         session = httpx.AsyncClient(base_url=url, headers=headers)
-        return AsyncClient(session=session)
+        return AsyncClient(session=session, serializer=serializer)
 
     session = httpx.Client(base_url=url, headers=headers)
-    return Client(session=session)
+    return Client(session=session, serializer=serializer)
 
 
 def new_client_from_environment(url: str = None):
@@ -632,27 +426,3 @@ def new_client_from_environment(url: str = None):
         )
 
     return Client(url=url, token=token)
-
-
-class OnePasswordConnectSDKError(RuntimeError):
-    pass
-
-
-class EnvironmentTokenNotSetException(OnePasswordConnectSDKError, TypeError):
-    pass
-
-
-class EnvironmentHostNotSetException(OnePasswordConnectSDKError, TypeError):
-    pass
-
-
-class FailedToRetrieveItemException(OnePasswordConnectSDKError):
-    pass
-
-
-class FailedToRetrieveVaultException(OnePasswordConnectSDKError):
-    pass
-
-
-class FailedToDeserializeException(OnePasswordConnectSDKError, TypeError):
-    pass
